% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nothing_profits.R
\name{do_nothing_profits}
\alias{do_nothing_profits}
\title{Calculate profits from a mine environment without taking any actions}
\usage{
do_nothing_profits(env)
}
\arguments{
\item{env}{list, a mining environment}
}
\value{
a list with data frames containing various profit metrics
}
\description{
Calculate profits from a mine environment without taking any actions
}
\examples{

do_nothing_profits(env1)

}
