% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gold_price.R
\name{gen_gold_price}
\alias{gen_gold_price}
\title{Generate gold price simulation values}
\usage{
gen_gold_price(rounds, years, gold_price_init, gold_price_delta)
}
\arguments{
\item{rounds}{integer, number of simulations}

\item{years}{integer, the number of years of data to generate (min 2)}

\item{gold_price_init}{int, the value of gold in the first year}

\item{gold_price_delta}{list, the parameters for the triangular distribution}
}
\value{
a data frame with dimensions _(rounds,years)_, with simulated gold
    price values
}
\description{
Generate gold price simulation values
}
\examples{
# Generate 10,000 simulations for 5 years:

gen_gold_price(10000, 5, 1700, list(a = -100, b = 500, c = 100))

}
