% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gold_prod.R
\name{gen_gold_prod}
\alias{gen_gold_prod}
\title{Generate gold production simulation values}
\usage{
gen_gold_prod(rounds, years, prod_init, prod_delta)
}
\arguments{
\item{rounds}{integer, number of simulations}

\item{years}{integer, the number of years of data to generate (min 2)}

\item{prod_init}{list, the parameters for a triangular distribution}

\item{prod_delta}{list, the parameters for a triangular distribution}
}
\value{
a data frame with dimensions _(rounds,years)_, with simulated gold
    production values
}
\description{
Generate gold production simulation values
}
\examples{
# Generate 10,000 simulations for 5 years:

gen_gold_prod(10000, 5, list(a = 300, b = 1400, c = 500),
                        list(a = 0.8, b = 1.8, c = 1.4))

}
