% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_independent.R
\name{gen_independent}
\alias{gen_independent}
\title{Generate a table of independent simulation values}
\usage{
gen_independent(rounds, years, t)
}
\arguments{
\item{rounds}{integer, number of simulations}

\item{years}{integer, the number of years of data to generate (min 2)}

\item{t}{list, the parameters for the triangular distribution}
}
\value{
a data frame with dimensions _(rounds,years)_, with simulated
    independent values given by **t**.
}
\description{
Generate a table of independent simulation values
}
\examples{
# Generate 10,000 simulations for 5 years:

gen_independent(10000, 5, list(a = 0.08, b = 0.16, c = 0.1275))

}
