% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_lease_profit.R
\name{gen_lease_profit}
\alias{gen_lease_profit}
\title{Generate leased mine profit simulation values}
\usage{
gen_lease_profit(rounds, years, profit_init, profit_delta)
}
\arguments{
\item{rounds}{integer, number of simulations}

\item{years}{integer, the number of years of data to generate (min 2)}

\item{profit_init}{list, the parameters for a triangular distribution}

\item{profit_delta}{list, the parameters for a triangular distribution}
}
\value{
a data frame with dimensions _(rounds,years)_, with simulated leased
    mine profit values
}
\description{
Generate leased mine profit simulation values
}
\examples{
# Generate 10,000 simulations for 5 years:

gen_lease_profit(10000, 5, list(a = 1000000, b = 3000000, c = 2000000),
                           list(a = 0.8, b = 1.45, c = 1.1))

}
