% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rep_mgmt.R
\name{gen_rep_mgmt}
\alias{gen_rep_mgmt}
\title{Generate reputation management cost simulation values}
\usage{
gen_rep_mgmt(rounds, years, fixed, rate, hours)
}
\arguments{
\item{rounds}{integer, number of simulations}

\item{years}{integer, the number of years of data to generate (min 2)}

\item{fixed}{int, fixed annual retainer for reputation management}

\item{rate}{int, fixed daily cost for reputation management work}

\item{hours}{list, the parameters for the triangular distribution}
}
\value{
a data frame with dimensions _(rounds,years)_, with simulated
    reputation management cost values
}
\description{
Generate reputation management cost simulation values
}
\examples{
# Generate 10,000 simulations for 5 years:

gen_rep_mgmt(10000, 5, 350000, 2500, list(a = 75, b = 350, c = 125))

}
