% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_selling_rate.R
\name{gen_selling_rate}
\alias{gen_selling_rate}
\title{Generate selling rate simulation values}
\usage{
gen_selling_rate(rounds, years, cost_init, cost_delta)
}
\arguments{
\item{rounds}{integer, number of simulations}

\item{years}{integer, the number of years of data to generate (min 2)}

\item{cost_init}{list, the parameters for a triangular distribution}

\item{cost_delta}{list, the parameters for a triangular distribution}
}
\value{
a data frame with dimensions _(rounds,years)_, with simulated selling
    cost rate values
}
\description{
Generate selling rate simulation values
}
\examples{
# Generate 10,000 simulations for 5 years:

gen_selling_rate(10000, 5, list(a = 0.05, b = 0.095, c = 0.07),
                           list(a = -0.1, b = 0.45, c = 0.2275))

}
