% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profits_with_actions.R
\name{profits_with_actions}
\alias{profits_with_actions}
\title{Calculate profits from a mine environment where actions are taken}
\usage{
profits_with_actions(env1, env3, params, sell_m2 = 0, find_m3 = 0)
}
\arguments{
\item{env1}{list, environment simulation for mines 1 and 2}

\item{env3}{list, environment simulation for mines 3}

\item{params}{list, mining simulation parameters}

\item{sell_m2}{integer, 1-5, the year in which the mine is sold}

\item{find_m3}{integer, 1-4, the year in which the third mine is found}
}
\value{
a list with data frames containing various profit metrics
}
\description{
Calculate profits from a mine environment where actions are taken
}
\examples{

profits_with_actions(env1, env3, 2, 3)

}
