% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profits_with_search.R
\name{profits_with_search}
\alias{profits_with_search}
\title{Calculate profits from a mine environment where actions are taken, and search
starts immediately}
\usage{
profits_with_search(env1, env3, params, sell_m2 = 0, search = 0)
}
\arguments{
\item{env1}{list, environment simulation for mines 1 and 2}

\item{env3}{list, environment simulation for mines 3}

\item{params}{list, mining simulation parameters}

\item{sell_m2}{integer, 1-5, the year in which the mine is sold}

\item{search}{integer, 0-6, the number of search parties to send}
}
\value{
a list with data frames containing various profit metrics
}
\description{
Calculate profits from a mine environment where actions are taken, and search
starts immediately
}
\examples{

profits_with_search(env1, env3, 2, 6)

}
