% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_sample.R
\name{triangle_sample}
\alias{triangle_sample}
\title{Generate random data from a triangular distribution}
\usage{
triangle_sample(n, a, b, c)
}
\arguments{
\item{n}{integer, number of random samples}

\item{a}{float, the minimum value for the distribution}

\item{b}{float, the maximum value for the distribution}

\item{c}{float, the most likely value for the distribution}
}
\value{
a vector of floats, drawn from the triangular distribution
}
\description{
Generate random data from a triangular distribution
}
\examples{
# Generate 10 samples from a triangular distribution with:
# - minimum value 1
# - maximum value 5
# - expected value 2.2

triangle_sample(10,1,5,2.2)

}
